開発: コメント読み上げ関連設定を新カテゴリ「コメント読み上げ」に集約
読み上げ設定・読み上げ停止設定・音声エンジンを個別のカテゴリやサブセクションから
新トップレベルカテゴリ「CommentSpeech（コメント読み上げ）」に集約。
SpeechEngineカテゴリを廃止し、旧プロトコルリンクのリダイレクト対応も追加。
