+++ app/components/settings/Settings.vue.ts
@@ -1,10 +1,10 @@
 import GenericFormGroups from 'components/obs/inputs/GenericFormGroups.vue';
 import { CategoryIcons } from 'components/settings/CategoryIcons';
 import CommentSettings from 'components/settings/CommentSettings.vue';
+import CommentSpeechSettings from 'components/settings/CommentSpeechSettings.vue';
 import ExtraSettings from 'components/settings/ExtraSettings.vue';
 import Hotkeys from 'components/settings/Hotkeys.vue';
 import LanguageSettings from 'components/settings/LanguageSettings.vue';
-import SpeechEngineSettings from 'components/settings/SpeechEngineSettings.vue';
 import SubStreamSettings from 'components/settings/SubStreamSettings.vue';
 import TranscriptionSettings from 'components/settings/TranscriptionSettings.vue';
 import ModalLayout from 'components/shared/ModalLayout.vue';
@@ -38,10 +38,11 @@ const CATEGORIES_WITH_TOC: string[] = [
   'Hotkeys',
   'Advanced',
   'Comment',
+  'CommentSpeech',
 ];
 
 // ニコニコログインが必要なカテゴリ
-const CATEGORIES_REQUIRING_LOGIN: SettingsCategory[] = ['Comment', 'SpeechEngine'];
+const CATEGORIES_REQUIRING_LOGIN: SettingsCategory[] = ['Comment', 'CommentSpeech'];
 
 @Component({
   components: {
@@ -53,7 +54,7 @@ const CATEGORIES_REQUIRING_LOGIN: SettingsCategory[] = ['Comment', 'SpeechEngine
     Hotkeys,
     LanguageSettings,
     CommentSettings,
-    SpeechEngineSettings,
+    CommentSpeechSettings,
     SubStreamSettings,
     TranscriptionSettings,
     TableOfContents,
