+++ app/components/settings/SoundDetectorSettings.vue
@@ -16,90 +16,90 @@
             </div>
           </div>
         </div>
+
+        <toc-section title="基本設定">
+          <div class="section">
+            <div class="input-label section-heading">
+              <label>基本設定</label>
+            </div>
+            <ObsListInput v-model="soundDetectorSourceModel" />
+            <p v-if="!sourceAvailable" class="source-unavailable-warning">
+              選択したソースがこのシーンに存在しません。入力音声ソースを選び直してください。
+            </p>
+            <ObsSliderInput v-model="soundThresholdDbModel" />
+          </div>
+        </toc-section>
+
+        <toc-section title="コメント読み上げテスト">
+          <div class="section">
+            <div class="input-label section-heading">
+              <label>コメント読み上げテスト</label>
+            </div>
+            <div v-if="soundDetectorSourceModel.value !== null">
+              <div class="sound-detector-volmeter-label">
+                音声検出:
+                <template v-if="soundDetectorEnabled">
+                  <span v-if="soundDetected === 'loud'">しゃべっています</span>
+                  <span v-else-if="soundDetected === 'no-signal'">無音(信号なし)</span>
+                  <span v-else>静か</span>
+                  <span v-if="sourceMuted">(ミュート状態)</span>
+                </template>
+                <span v-else>設定がOFFです</span>
+              </div>
+              <div class="sound-detector-volmeter-container">
+                <SoundDetectorVolmeter
+                  v-for="audioSource in soundDetectorAudioSources"
+                  :audioSource="audioSource"
+                  :threshold="soundThresholdDbModel.value"
+                  :enabled="soundDetectorEnabled"
+                  :key="audioSource.sourceId"
+                />
+              </div>
+
+              <button
+                v-if="!isTestPlaybackActive"
+                data-color="secondary"
+                data-size="md"
+                data-variant="light"
+                data-radius="sm"
+                class="basic-button"
+                :disabled="!synthesizerEnabled || !soundDetectorEnabled"
+                @click="startContinuousPlayback"
+              >
+                テストを開始
+              </button>
+              <button
+                v-else
+                data-color="primary"
+                data-size="md"
+                data-variant="filled"
+                data-radius="sm"
+                class="basic-button"
+                :disabled="!soundDetectorEnabled"
+                @click="stopContinuousPlayback"
+              >
+                テストを停止
+              </button>
+
+              <p class="section-notice-text">テスト用の視聴者コメントを読み上げます。スピーカーで鳴らした音声がマイクに入らないよう注意してください</p>
+
+            </div>
+          </div>
+        </toc-section>
+
+        <toc-section title="詳細設定">
+          <div class="section">
+            <div class="input-label section-heading--dropdown" :class="{ 'is-collapsed': collapsed }" @click="collapsed = !collapsed">
+              <label>詳細設定</label>
+              <i :class="collapsed ? 'icon-arrow-bottom-fill' : 'icon-arrow-top-fill'" />
+            </div>
+            <div v-if="!collapsed">
+              <ObsSliderInput v-model="resumeSilenceMsModel" />
+              <ObsListInput v-model="soundDetectedSpeechActionModel" />
+            </div>
+          </div>
+        </toc-section>
       </div>
-
-      <toc-section title="基本設定">
-        <div class="section">
-          <div class="input-label section-heading">
-            <label>基本設定</label>
-          </div>
-          <ObsListInput v-model="soundDetectorSourceModel" />
-          <p v-if="!sourceAvailable" class="source-unavailable-warning">
-            選択したソースがこのシーンに存在しません。入力音声ソースを選び直してください。
-          </p>
-          <ObsSliderInput v-model="soundThresholdDbModel" />
-        </div>
-      </toc-section>
-
-      <toc-section title="コメント読み上げテスト">
-        <div class="section">
-          <div class="input-label section-heading">
-            <label>コメント読み上げテスト</label>
-          </div>
-          <div v-if="soundDetectorSourceModel.value !== null">
-            <div class="sound-detector-volmeter-label">
-              音声検出:
-              <template v-if="soundDetectorEnabled">
-                <span v-if="soundDetected === 'loud'">しゃべっています</span>
-                <span v-else-if="soundDetected === 'no-signal'">無音(信号なし)</span>
-                <span v-else>静か</span>
-                <span v-if="sourceMuted">(ミュート状態)</span>
-              </template>
-              <span v-else>設定がOFFです</span>
-            </div>
-            <div class="sound-detector-volmeter-container">
-              <SoundDetectorVolmeter
-                v-for="audioSource in soundDetectorAudioSources"
-                :audioSource="audioSource"
-                :threshold="soundThresholdDbModel.value"
-                :enabled="soundDetectorEnabled"
-                :key="audioSource.sourceId"
-              />
-            </div>
-
-            <button
-              v-if="!isTestPlaybackActive"
-              data-color="secondary"
-              data-size="md"
-              data-variant="light"
-              data-radius="sm"
-              class="basic-button"
-              :disabled="!synthesizerEnabled || !soundDetectorEnabled"
-              @click="startContinuousPlayback"
-            >
-              テストを開始
-            </button>
-            <button
-              v-else
-              data-color="primary"
-              data-size="md"
-              data-variant="filled"
-              data-radius="sm"
-              class="basic-button"
-              :disabled="!soundDetectorEnabled"
-              @click="stopContinuousPlayback"
-            >
-              テストを停止
-            </button>
-
-            <p class="section-notice-text">テスト用の視聴者コメントを読み上げます。スピーカーで鳴らした音声がマイクに入らないよう注意してください</p>
-
-          </div>
-        </div>
-      </toc-section>
-
-      <toc-section title="詳細設定">
-        <div class="section">
-          <div class="input-label section-heading--dropdown" :class="{ 'is-collapsed': collapsed }" @click="collapsed = !collapsed">
-            <label>詳細設定</label>
-            <i :class="collapsed ? 'icon-arrow-bottom-fill' : 'icon-arrow-top-fill'" />
-          </div>
-          <div v-if="!collapsed">
-            <ObsSliderInput v-model="resumeSilenceMsModel" />
-            <ObsListInput v-model="soundDetectedSpeechActionModel" />
-          </div>
-        </div>
-      </toc-section>
     </toc-section>
   </div>
 </template>
